Rename `factories` -> `factory`
